Add BaseExecutor & refactor executor module
