+++ awswrangler/s3/_select.py
@@ -13,7 +13,7 @@ import pyarrow as pa
 
 from awswrangler import _data_types, _utils, exceptions
 from awswrangler._distributed import engine
-from awswrangler._threading import _get_executor, _ThreadPoolExecutor
+from awswrangler._executor import _BaseExecutor, _get_executor
 from awswrangler.distributed.ray import ray_get
 from awswrangler.s3._describe import size_objects
 from awswrangler.s3._list import _path2list
@@ -81,7 +81,7 @@ def _select_object_content(
 @engine.dispatch_on_engine
 def _select_query(
     path: str,
-    executor: _ThreadPoolExecutor,
+    executor: _BaseExecutor,
     sql: str,
     input_serialization: str,
     input_serialization_params: Dict[str, Union[bool, str]],
@@ -282,7 +282,7 @@ def select_query(
     _logger.debug("kwargs:\n%s", pprint.pformat(select_kwargs))
 
     arrow_kwargs = _data_types.pyarrow2pandas_defaults(use_threads=use_threads, kwargs=pyarrow_additional_kwargs)
-    executor = _get_executor(use_threads=use_threads)
+    executor: _BaseExecutor = _get_executor(use_threads=use_threads)
     tables = list(
         itertools.chain(*ray_get([_select_query(path=path, executor=executor, **select_kwargs) for path in paths]))
     )
